+++ src/components/NavBar/navItems.ts
@@ -17,6 +17,10 @@ const NAV_ITEMS: Array<NavItem> = [
         label: "one handed scrambles",
         href: "tools/ohscramble/",
       },
+      {
+        label: "test page",
+        href: "tools/testpage/",
+      },
     ],
   },
 ]
