+++ src/libv2/puzzles/index.ts
@@ -1,0 +1,1 @@
+export * from "./cube3x3";
